requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
bcrypt
python-dotenv
email-validator
